+++ apps/backyard/src/app/current-data/dashboard/create-round.tsx
@@ -1,24 +1,19 @@
+import { useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Card } from 'ui';
+import { Button, Card } from 'ui';
 
-// TODO: move to ui pacakge
-import Button from '@/components/button';
 import { useRounds } from '@/hooks/current-data/use-rounds';
 import { notify } from '@/utils/notify';
 
-export default function RoundCreateView() {
+export default function CreateRoundView() {
   const navigate = useNavigate();
   const { rounds, createRound } = useRounds();
 
-  const nr =
-    rounds.reduce(
-      (nextNr, round) => (round.nr > nextNr ? round.nr : nextNr),
-      0
-    ) + 1;
+  const nr = useRef((rounds.at(-1)?.nr || 0) + 1);
 
-  const create = () => {
-    notify(createRound(nr), `Runde ${nr} angelegt`);
-    navigate('../runden');
+  const create = async () => {
+    await notify(createRound(nr.current), `Runde ${nr.current} angelegt`);
+    navigate('../spiele');
   };
 
   return (
@@ -26,7 +21,7 @@ export default function RoundCreateView() {
       <Card>
         <Card.Header>Neue Runde</Card.Header>
         <div className="flex p-4 items-center justify-between">
-          <h2 className="text-lg pl-2 font-semibold">Runde {nr}</h2>
+          <h2 className="text-lg pl-2 font-semibold">Runde {nr.current}</h2>
           <Button primary onClick={create}>
             Anlegen
           </Button>