+++ apps/backyard/src/app/current-data/dashboard/index.tsx
@@ -21,6 +21,30 @@ const items: {
   visible: (championship: Championship | undefined, rounds: Round[]) => boolean;
 }[] = [
   {
+    title: 'Ergebnisse eintragen',
+    description: 'Spiel-Ergebnisse eintragen und auswerten',
+    icon: ScaleIcon,
+    background: 'bg-purple-500',
+    route: './ergebnisse-eintragen',
+    visible: (championship) => !!championship && false,
+  },
+  {
+    title: 'Tipps eintragen',
+    description: 'Tipps der Mitspieler erfassen.',
+    icon: PencilSquareIcon,
+    background: 'bg-indigo-500',
+    route: './tipps',
+    visible: (championship) => !!championship && false,
+  },
+  {
+    title: 'Spielansetzungen',
+    description: 'Bearbeite die Ansetzungen einer Runde',
+    icon: MegaphoneIcon,
+    background: 'bg-blue-500',
+    route: './spiele',
+    visible: (championship, rounds) => !!championship && rounds.length > 0,
+  },
+  {
     title: 'Neue Runde',
     description: 'Lege eine neue (Monats-) Runde an',
     icon: CalendarIcon,
@@ -35,30 +59,6 @@ const items: {
     background: 'bg-pink-500',
     route: './neues-turnier',
     visible: () => true,
-  },
-  {
-    title: 'Spielansetzungen',
-    description: 'Bearbeite die Ansetzungen einer Runde',
-    icon: MegaphoneIcon,
-    background: 'bg-blue-500',
-    route: './runden',
-    visible: (championship, rounds) => !!championship && rounds.length > 0,
-  },
-  {
-    title: 'Tipps eintragen',
-    description: 'Tipps der Mitspieler erfassen.',
-    icon: PencilSquareIcon,
-    background: 'bg-indigo-500',
-    route: './tipps',
-    visible: (championship) => !!championship && false,
-  },
-  {
-    title: 'Ergebnisse eintragen',
-    description: 'Spiel-Ergebnisse eintragen und auswerten',
-    icon: ScaleIcon,
-    background: 'bg-purple-500',
-    route: './ergebnisse-eintragen',
-    visible: (championship) => !!championship && false,
   },
 ];