+++ apps/backyard/src/hooks/current-data/use-rounds.ts
@@ -8,17 +8,13 @@ export function useRounds() {
   const { currentChampionship } = useCurrentChampionship();
   const rounds = useRecoilValue(roundsState(currentChampionship?.id));
 
-  const createRound = (nr: number) => {
-    return createEntity<Round>(
-      `championships/${currentChampionship?.id}/rounds`,
-      {
-        id: `runde-${nr}`,
-        nr,
-        published: false,
-        completed: false,
-      }
-    );
-  };
+  const createRound = async (nr: number) =>
+    createEntity<Round>(`championships/${currentChampionship?.id}/rounds`, {
+      id: `runde-${nr}`,
+      nr,
+      published: false,
+      completed: false,
+    });
 
   return { rounds, createRound };
 }