+++ packages/lib/src/model/match.ts
@@ -1,13 +1,12 @@
-import { Timestamp } from 'firebase/firestore';
 import { BaseModel } from './base/model';
 
 export type Match = BaseModel & {
   roundId: string;
   nr: number;
-  date: Timestamp;
-  league: string;
-  hometeam: string;
-  awayteam: string;
+  date: string;
+  leagueId: string;
+  hometeamId: string;
+  awayteamId: string;
   result: string;
   points: number;
 };