+++ packages/tailwind-config/index.d.ts
@@ -1,0 +1,5 @@
+import { Config } from 'tailwindcss';
+
+declare module 'tailwind-config/backyard.config.cjs' {
+  export = Config;
+}