+++ packages/ui/src/components/container/card.tsx
@@ -5,16 +5,12 @@ type CardProps = {
 };
 
 export function Card({ children }: CardProps) {
-  return (
-    <div className="divide-y divide-gray-200 overflow-hidden rounded-lg bg-white shadow">
-      {children}
-    </div>
-  );
+  return <div className="rounded-lg bg-white shadow">{children}</div>;
 }
 
 function Header({ children }: CardProps) {
   return (
-    <div className="bg-white px-4 py-5 sm:px-6">
+    <div className="px-4 py-5 sm:px-6 border-b border-gray-200">
       <h3 className="text-lg font-medium leading-6 text-gray-900">
         {children}
       </h3>